Rename : img파일 src/assets로 이동
